+++ libs/frontend/modules/element/src/graphql/index.ts
@@ -1,2 +1,2 @@
-export * from './element.endpoints.v2-extended'
+export * from './element.endpoints.v2.graphql.gen'
 export * from './Element.fragment.v2.graphql.gen'
