--- libs/frontend/modules/element/src/store/elementEndpoints.v2.ts
@@ -1,162 +1,0 @@
-import { RootState } from '@reduxjs/toolkit/dist/query/core/apiState'
-import { Recipe } from '@reduxjs/toolkit/dist/query/core/buildThunks'
-import { merge, pickBy } from 'lodash'
-import { DefaultRootState } from 'react-redux'
-import {
-  api as generatedApi,
-  ElementEdgeFragment,
-  ElementFragment,
-  GetElementsWithGraphQuery,
-  normalizeVertices,
-} from '../graphql'
-
-type GraphUpdateHandler = Recipe<GetElementsWithGraphQuery>
-
-const fulfilledRequests: Array<string> = []
-
-const getGraphEntry = (currentGraphRootId: string) => ({
-  variables: { where: { id: currentGraphRootId } },
-})
-
-const createEdges = (
-  created: Array<ElementFragment>,
-): Array<ElementEdgeFragment> =>
-  created
-    .filter((x) => x.parentElement)
-    .map(({ id, parentElement, parentElementConnection }) => ({
-      // parentElement is defined because of filter
-      source: (parentElement as ElementFragment).id,
-      target: id,
-      order: parentElementConnection.edges[0]?.order,
-    }))
-
-const removeEdges = (edges: Array<ElementEdgeFragment>, ids: Array<string>) =>
-  edges.filter((x) => !ids.includes(x.source) && !ids.includes(x.target))
-
-const onCreate =
-  (rootId: string, created: Array<ElementFragment>): GraphUpdateHandler =>
-  (draft) => {
-    const oldGraph = draft.elements[rootId]
-    draft.elements[rootId] = {
-      vertices: merge(oldGraph.vertices, normalizeVertices(created)),
-      edges: oldGraph.edges.concat(createEdges(created)),
-    }
-  }
-
-const onUpdate =
-  (rootId: string, updated: Array<ElementFragment>): GraphUpdateHandler =>
-  (draft) => {
-    const oldGraph = draft.elements[rootId]
-    const updatedIds = updated.map((x) => x.id)
-
-    draft.elements[rootId] = {
-      vertices: merge(oldGraph.vertices, normalizeVertices(updated)),
-      edges: removeEdges(oldGraph.edges, updatedIds).concat(
-        createEdges(updated),
-      ),
-    }
-  }
-
-const onDelete =
-  (rootId: string, deletedIds: Array<string>): GraphUpdateHandler =>
-  (draft) => {
-    const oldGraph = draft.elements[rootId]
-    draft.elements[rootId] = {
-      vertices: pickBy(
-        oldGraph.vertices,
-        (value, key) => !deletedIds.includes(key),
-      ),
-      edges: removeEdges(oldGraph.edges, deletedIds),
-    }
-  }
-
-const updateCache = (rootId: string, updateRecipe: GraphUpdateHandler) =>
-  generatedApi.util.updateQueryData(
-    'GetElementsWithGraph',
-    getGraphEntry(rootId),
-    updateRecipe,
-  )
-
-const getGraphRootId = (getState: () => RootState<any, any, 'api'>) =>
-  (getState() as unknown as DefaultRootState).element.currentGraphRootId
-
-const runGuards = (
-  requestId: string,
-  getState: () => RootState<any, any, 'api'>,
-  callback: (rootId: string) => void,
-) => {
-  if (fulfilledRequests.includes(requestId)) {
-    return
-  }
-
-  const currentGraphRootId = getGraphRootId(getState)
-
-  if (!currentGraphRootId) {
-    console.log('Graph cache is not updated, Graph root is not provided')
-
-    return
-  }
-
-  callback(currentGraphRootId)
-  fulfilledRequests.push(requestId)
-}
-
-export const elementApiV2 = generatedApi.enhanceEndpoints({
-  endpoints: {
-    CreateElements: {
-      async onQueryStarted(input, api) {
-        const { dispatch, queryFulfilled, getState, requestId } = api
-        const { data } = await queryFulfilled
-        runGuards(requestId, getState, async (rootId) => {
-          const createdElements = data.createElements.elements
-          dispatch(updateCache(rootId, onCreate(rootId, createdElements)))
-        })
-      },
-    },
-    DuplicateElement: {
-      async onQueryStarted(input, api) {
-        const { dispatch, queryFulfilled, getState, requestId } = api
-        const { data } = await queryFulfilled
-        runGuards(requestId, getState, async (rootId) => {
-          const duplicatedElements = data.duplicateElement.elements
-          dispatch(updateCache(rootId, onCreate(rootId, duplicatedElements)))
-        })
-      },
-    },
-    UpdateElements: {
-      async onQueryStarted(input, api) {
-        const { dispatch, queryFulfilled, getState, requestId } = api
-        const { data } = await queryFulfilled
-        runGuards(requestId, getState, async (rootId) => {
-          const updatedElements = data.updateElements.elements
-          dispatch(updateCache(rootId, onUpdate(rootId, updatedElements)))
-        })
-      },
-    },
-    DeleteElements: {
-      async onQueryStarted(input, api) {
-        const { dispatch, queryFulfilled, getState, requestId } = api
-        await queryFulfilled
-        runGuards(requestId, getState, async (rootId) => {
-          const deletedIds =
-            (input?.variables?.where?.id_IN as Array<string>) || []
-
-          dispatch(updateCache(rootId, onDelete(rootId, deletedIds)))
-        })
-      },
-    },
-  },
-})
-
-export { generatedApi as elementEndpointsV2 }
-
-export const {
-  useCreateElementsMutation,
-  useDeleteElementsMutation,
-  useUpdateElementsMutation,
-  useGetElementsQuery,
-  useLazyGetElementsQuery,
-  useGetElementsWithGraphQuery,
-  useLazyGetElementsWithGraphQuery,
-  useDuplicateElementMutation,
-} = elementApiV2
