+++ libs/frontend/modules/element/src/store/elementEndpoints.v2/index.ts
@@ -1,0 +1,1 @@
+export * from './elementEndpoints.v2'
